Rename VictoryTheme
