+++ packages/component-library/src/_Themes/index.js
@@ -1,7 +1,7 @@
 import Brand from "./BrandTheme";
-import DataVisualization from "./DataVisualizationTheme";
+import Victory from "./VictoryTheme";
 import CrazyVisualization from "./CrazyVisualizationTheme";
 
 export const BrandTheme = Brand;
-export const DataVisualizationTheme = DataVisualization;
+export const VictoryTheme = Victory;
 export const CrazyVisualizationTheme = CrazyVisualization;
